+++ django_typer/tests/apps/test_app/management/commands/empty.py
@@ -1,0 +1,5 @@
+from django_typer import TyperCommand
+
+
+class Command(TyperCommand):
+    pass