+++ django_typer/tests/apps/test_app/management/commands/help_precedence2.py
@@ -6,6 +6,10 @@ from django_typer import TyperCommand, command, initialize
 
 
 class Command(TyperCommand):
+    """
+    Class docstring.
+    """
+
     help = _("Test minimal TyperCommand subclass - class member")
 
     @initialize(help=_("Test minimal TyperCommand subclass - callback method"))