+++ django_typer/tests/apps/test_app/management/commands/help_precedence3.py
@@ -6,6 +6,10 @@ from django_typer import TyperCommand, command, initialize
 
 
 class Command(TyperCommand):
+    """
+    Class docstring.
+    """
+
     @initialize(help=_("Test minimal TyperCommand subclass - callback method"))
     def init(self, verbosity: int = 1):
         """