+++ django_typer/tests/apps/test_app/management/commands/help_precedence5.py
@@ -6,6 +6,10 @@ from django_typer import TyperCommand, command
 
 
 class Command(TyperCommand):
+    """
+    Class docstring.
+    """
+
     @command(help=_("Test minimal TyperCommand subclass - command method"))
     def handle(self, arg1: str, arg2: str, arg3: float = 0.5, arg4: int = 1):
         """