+++ django_typer/tests/apps/test_app/management/commands/help_precedence6.py
@@ -5,6 +5,10 @@ from django_typer import TyperCommand
 
 
 class Command(TyperCommand):
+    """
+    Class docstring.
+    """
+
     def handle(self, arg1: str, arg2: str, arg3: float = 0.5, arg4: int = 1):
         """
         Test minimal TyperCommand subclass - docstring