+++ django_typer/tests/apps/test_app/management/commands/help_precedence8.py
@@ -6,6 +6,10 @@ from django_typer import TyperCommand
 
 
 class Command(TyperCommand, help=_("Test minimal TyperCommand subclass - typer param")):
+    """
+    Class docstring.
+    """
+
     help = _("Test minimal TyperCommand subclass - class member")
 
     def handle(self, arg1: str, arg2: str, arg3: float = 0.5, arg4: int = 1):